Moved templatetags and dependencies inside phonenumber plugin
