+++ two_factor/plugins/phonenumber/models.py
@@ -10,6 +10,8 @@ from phonenumber_field.modelfields import PhoneNumberField
 
 from two_factor.gateways import make_call, send_sms
 
+from .utils import format_phone_number, mask_phone_number
+
 PHONE_METHODS = (
     ('call', _('Phone Call')),
     ('sms', _('Text Message')),
@@ -73,6 +75,14 @@ class PhoneDevice(ThrottlingMixin, Device):
             make_call(device=self, token=token)
         else:
             send_sms(device=self, token=token)
+
+    @property
+    def generate_challenge_button_title(self):
+        number = mask_phone_number(format_phone_number(self.number))
+        if self.method == 'sms':
+            return _('Send text message to %s') % number
+        else:
+            return _('Call number %s') % number
 
     def get_throttle_factor(self):
         return getattr(settings, 'TWO_FACTOR_PHONE_THROTTLE_FACTOR', 1)
