+++ two_factor/templates/two_factor/core/login.html
@@ -1,5 +1,5 @@
 {% extends "two_factor/_base_focus.html" %}
-{% load i18n two_factor %}
+{% load i18n %}
 
 {% block content %}
   <h1>{% block title %}{% trans "Login" %}{% endblock %}</h1>
@@ -35,7 +35,7 @@
         {% for other in other_devices %}
         <button name="challenge_device" value="{{ other.persistent_id }}"
                 class="btn btn-secondary btn-block" type="submit">
-          {{ other|device_action }}
+          {{ other.generate_challenge_button_title }}
         </button>
       {% endfor %}</p>
     {% endif %}
