+++ two_factor/templates/two_factor/profile/profile.html
@@ -1,5 +1,5 @@
 {% extends "two_factor/_base.html" %}
-{% load i18n two_factor %}
+{% load i18n %}
 
 {% block content %}
   <h1>{% block title %}{% trans "Account Security" %}{% endblock %}</h1>
@@ -8,7 +8,7 @@
     {% if default_device_type == 'TOTPDevice' %}
       <p>{% trans "Tokens will be generated by your token generator." %}</p>
     {% elif default_device_type == 'PhoneDevice' %}
-      <p>{% blocktrans with primary=default_device|device_action %}Primary method: {{ primary }}{% endblocktrans %}</p>
+      <p>{% blocktrans with primary=default_device.generate_challenge_button_title %}Primary method: {{ primary }}{% endblocktrans %}</p>
     {% elif default_device_type == 'RemoteYubikeyDevice' %}
       <p>{% blocktrans %}Tokens will be generated by your YubiKey.{% endblocktrans %}</p>
     {% endif %}
@@ -20,9 +20,9 @@
       <ul>
         {% for phone in backup_phones %}
           <li>
-            {{ phone|device_action }}
+            {{ phone.generate_challenge_button_title }}
             <form method="post" action="{% url 'two_factor:phone_delete' phone.id %}"
-                  onsubmit="return confirm('Are you sure?')">
+                  onsubmit="return confirm({% trans 'Are you sure?' %})">
               {% csrf_token %}
               <button class="btn btn-sm btn-warning"
                       type="submit">{% trans "Unregister" %}</button>
